requirements.txt
faiss-cpu==1.7.4
numpy>=1.19.0
sentence-transformers>=2.2.2
requests>=2.31.0
beautifulsoup4>=4.12.2
tqdm>=4.66.1 